Try out pydantic2zod (unsuccessfully)
This experiment didn't work either. It doesn't end in error, it must
basically emits nothing, so it either fails to find any models to
convert or it silently errors out.

This is the total output:

```typescript
/**
 * NOTE: automatically generated by the pydantic2zod compiler.
 */

import { z } from "zod";
```

So this tool, too, either doesn't work or will require more massaging to
make it work. On the plus side, it seems more thoughtfully-designed than
pydantic-to-typescript.